Migrate fjs/cli to .f.mjs, split type-level API into types.ts
proof.f.ts stays TypeScript for now: it has a runtime dependency on
effects/node/virtual/module.f.ts, still unmigrated. Updates the
cas/cli and top-level module.f.ts dependents, plus stale module.f.ts
doc references (including line-number citations) across README.md and
five todo docs.
