+++ fjs/cas/cli/module.f.ts
@@ -8,7 +8,8 @@ import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { forEachStep, pure, step } from '../../effects/module.f.mjs'
 import { errorExit, log, writeFromStream } from '../../effects/node/module.f.mjs'
 import type { All, Read, Write, WriteFile } from '../../effects/node/types.ts'
-import { dispatch, type Commands } from '../../cli/module.f.ts'
+import { dispatch } from '../../cli/module.f.mjs'
+import type { Commands } from '../../cli/types.ts'
 import type { MemOp } from '../../effects/memory/types.ts'
 import { casAddFile, fileCas, type FileCasOperation } from '../module.f.ts'
