+++ fjs/cli/proof.f.ts
@@ -1,7 +1,8 @@
 import { pure } from '../effects/module.f.mjs'
 import type { NodeOp, NodeProgramOptions } from '../effects/node/types.ts'
 import { defaultNodeProgramOptions, emptyState, virtual } from '../effects/node/virtual/module.f.ts'
-import { dispatch, type Commands } from './module.f.ts'
+import { dispatch } from './module.f.mjs'
+import type { Commands } from './types.ts'
 import { assert, assertEq } from '../asserts/module.f.mjs'
 
 const makeOptions = (args: readonly string[]): NodeProgramOptions =>
