+++ fjs/effects/todo/map-step-combinator.md
@@ -39,7 +39,7 @@ Also `fjs/dev/module.f.ts`, `fjs/cas/evo/module.f.ts`,
 yield an exit code" shape of a `NodeProgram`:
 
 ```ts
-// fjs/cli/module.f.ts:38-40
+// fjs/cli/module.f.mjs:39-41
 return step(
     log(helpText),
     () => pure(0))
