+++ fjs/module.f.ts
@@ -9,7 +9,8 @@ import { commands as casCommands } from './cas/cli/module.f.ts'
 import { main as ciMain } from './ci/module.f.ts'
 import { import_ } from './effects/node/module.f.mjs'
 import type { NodeOp, NodeProgram } from './effects/node/types.ts'
-import { dispatch, type Commands } from './cli/module.f.ts'
+import { dispatch } from './cli/module.f.mjs'
+import type { Commands } from './cli/types.ts'
 import { casMcpServer } from './mcp/module.f.ts'
 import { pure, step } from './effects/module.f.mjs'
 import { unwrap } from './types/result/module.f.mjs'
